+++ tests/lantern_tests/models/item/base/test_item_base.py
@@ -146,7 +146,7 @@ class TestItemCore:
             (False, [], AccessLevel.NONE),
             (True, [], AccessLevel.NONE),
             (True, [Permission(directory="~nerc", group="~bas-staff")], AccessLevel.BAS_STAFF),
-            (True, [Permission(directory="*", group="*")], AccessLevel.PUBLIC),
+            (True, [Permission(directory="*", group="*")], AccessLevel.OPEN_ACCESS),
             (True, [Permission(directory="x", group="x"), Permission(directory="y", group="y")], AccessLevel.UNKNOWN),
         ],
     )
@@ -707,7 +707,7 @@ class TestItemSummaryBase:
             (AccessLevel.NONE, True),
             (AccessLevel.UNKNOWN, True),
             (AccessLevel.BAS_STAFF, True),
-            (AccessLevel.PUBLIC, False),
+            (AccessLevel.OPEN_ACCESS, False),
         ],
     )
     def test_restricted(