+++ tests/resources/records/item_cat_product_restricted_bas_staff.py
@@ -1,0 +1,116 @@
+from bas_metadata_library.standards.magic_administration.v1 import AdministrationMetadata
+
+from lantern.lib.metadata_library.models.record.elements.common import (
+    Address,
+    Constraints,
+    Contact,
+    ContactIdentity,
+    OnlineResource,
+)
+from lantern.lib.metadata_library.models.record.elements.distribution import (
+    Distribution,
+    Distributions,
+    Format,
+    Size,
+    TransferOption,
+)
+from lantern.lib.metadata_library.models.record.enums import (
+    ContactRoleCode,
+    HierarchyLevelCode,
+    OnlineResourceFunctionCode,
+)
+from lantern.lib.metadata_library.models.record.presets.admin import BAS_STAFF as BAS_STAFF_PERMISSION
+from lantern.lib.metadata_library.models.record.presets.constraints import BAS_STAFF, MAGIC_PRODUCTS_V1
+from lantern.lib.metadata_library.models.record.utils.admin import set_admin
+from tests.resources.admin_keys import test_keys
+from tests.resources.records.utils import make_record, relate_products
+
+# A restricted record for testing a catalogue item restricted to BAS Staff.
+
+record = make_record(
+    open_access=False,
+    file_identifier="57327327-4623-4247-af86-77fb43b7f45b",
+    hierarchy_level=HierarchyLevelCode.PRODUCT,
+    title="Test Resource - Product restricted to BAS Staff",
+    abstract="Item to test a Product with a BAS Staff restricted access constraint is presented correctly.",
+)
+# add related peers
+record.identification.aggregations.extend(relate_products(record.file_identifier))
+
+# change access and licence
+record.identification.constraints = Constraints([BAS_STAFF, MAGIC_PRODUCTS_V1])
+# add admin metadata to reflect access
+keys = test_keys()
+admin = AdministrationMetadata(id=record.file_identifier, resource_permissions=[BAS_STAFF_PERMISSION])
+set_admin(keys=keys, record=record, admin_meta=admin)
+
+# add example distribution to test restricted state handling
+distributor = Contact(
+    organisation=ContactIdentity(
+        name="Mapping and Geographic Information Centre, British Antarctic Survey",
+        href="https://ror.org/01rhff309",
+        title="ror",
+    ),
+    phone="+44 (0)1223 221400",
+    email="magic@bas.ac.uk",
+    address=Address(
+        delivery_point="British Antarctic Survey, High Cross, Madingley Road",
+        city="Cambridge",
+        administrative_area="Cambridgeshire",
+        postal_code="CB3 0ET",
+        country="United Kingdom",
+    ),
+    online_resource=OnlineResource(
+        href="https://www.bas.ac.uk/teams/magic",
+        title="Mapping and Geographic Information Centre (MAGIC) - BAS public website",
+        description="General information about the BAS Mapping and Geographic Information Centre (MAGIC) from the British Antarctic Survey (BAS) public website.",
+        function=OnlineResourceFunctionCode.INFORMATION,
+    ),
+    role={ContactRoleCode.DISTRIBUTOR},
+)
+record.distribution = Distributions(
+    [
+        Distribution(
+            distributor=distributor,
+            format=Format(
+                format="GeoJSON",
+                href="https://www.iana.org/assignments/media-types/application/geo+json",
+            ),
+            transfer_option=TransferOption(
+                size=Size(unit="bytes", magnitude=24 * 1024 * 1024),
+                online_resource=OnlineResource(
+                    href="x",
+                    function=OnlineResourceFunctionCode.DOWNLOAD,
+                    title="GeoJSON",
+                    description="Access information as a GeoJSON file.",
+                ),
+            ),
+        ),
+        Distribution(
+            distributor=distributor,
+            format=Format(
+                format="PDF",
+                href="https://www.iana.org/assignments/media-types/application/pdf",
+            ),
+            transfer_option=TransferOption(
+                size=Size(unit="bytes", magnitude=321 * 1024 * 1024),
+                online_resource=OnlineResource(
+                    href="x",
+                    function=OnlineResourceFunctionCode.DOWNLOAD,
+                    title="PDF",
+                    description="Access information as a PDF file.",
+                ),
+            ),
+        ),
+        Distribution(
+            distributor=distributor,
+            transfer_option=TransferOption(
+                online_resource=OnlineResource(
+                    href="sftp://san.nerc-bas.ac.uk/data/x",
+                    function=OnlineResourceFunctionCode.DOWNLOAD,
+                    title="Access from the BAS SAN",
+                ),
+            ),
+        ),
+    ]
+)